@@ -8,9 +8,10 @@
 		"lint": "vue-cli-service lint"
 	},
 	"dependencies": {
-		"@vue/composition-api": "^1.7.0",
 		"core-js": "^2.6.5",
-		"vue": "^2.6.14"
+		"vue": "^2.7.1",
+		"vuex": "^3.6.2",
+		"vuexok": "^1.0.2"
 	},
 	"devDependencies": {
 		"@babel/core": "^7.4.3",
@@ -23,8 +24,9 @@
 		"babel-eslint": "^10.0.1",
 		"ts-loader": "^9.3.1",
 		"typescript": "^4.7.4",
-		"eslint": "^5.8.0",
-		"eslint-plugin-vue": "^5.0.0",
+		"eslint": "^6.7.2",
+		"eslint-plugin-import": "^2.20.2",
+		"eslint-plugin-vue": "^6.2.2",
 		"vue-template-compiler": "~2.6.14"
 	},
 	"eslintConfig": {
